sharding/requirements.txt
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
